Inline cordova/builder into cordova/modulemapper
